Report error for assignment of private class members.

+++ typescript/file_analyzer/src/analyzer/assign/class.rs
@@ -7,6 +7,7 @@ use stc_ts_types::Class;
 use stc_ts_types::ClassMember;
 use stc_ts_types::Type;
 use swc_common::EqIgnoreSpan;
+use swc_ecma_ast::Accessibility;
 
 impl Analyzer<'_, '_> {
     pub(super) fn assign_to_class(&mut self, opts: AssignOpts, l: &Class, r: &Type) -> ValidationResult<()> {
@@ -98,6 +99,8 @@ impl Analyzer<'_, '_> {
         l: &ClassMember,
         r: &[ClassMember],
     ) -> ValidationResult<()> {
+        let span = opts.span;
+
         match l {
             ClassMember::Constructor(lc) => {
                 for rm in r {
@@ -111,12 +114,20 @@ impl Analyzer<'_, '_> {
                 }
             }
             ClassMember::Method(lm) => {
+                if lm.accessibility == Some(Accessibility::Private) {
+                    return Err(box Error::PrivateMethodIsDifferent { span });
+                }
+
                 for rmember in r {
                     match rmember {
                         ClassMember::Constructor(_) => {}
                         ClassMember::Method(rm) => {
                             //
                             if self.assign(&lm.key.ty(), &rm.key.ty(), opts.span).is_ok() {
+                                if rm.accessibility == Some(Accessibility::Private) {
+                                    return Err(box Error::PrivateMethodIsDifferent { span });
+                                }
+
                                 // TODO: Parameters.
                                 self.assign_with_opts(opts, &lm.ret_ty, &rm.ret_ty)
                                     .context("tried to assign return type of a class method")?;
@@ -134,12 +145,20 @@ impl Analyzer<'_, '_> {
                 }
             }
             ClassMember::Property(lp) => {
+                if lp.accessibility == Some(Accessibility::Private) {
+                    return Err(box Error::PrivatePropertyIsDifferent { span });
+                }
+
                 for rm in r {
                     match rm {
                         ClassMember::Constructor(_) => {}
                         ClassMember::Method(_) => {}
                         ClassMember::Property(rp) => {
                             if self.assign(&lp.key.ty(), &rp.key.ty(), opts.span).is_ok() {
+                                if rp.accessibility == Some(Accessibility::Private) {
+                                    return Err(box Error::PrivatePropertyIsDifferent { span });
+                                }
+
                                 if let Some(lt) = &lp.value {
                                     if let Some(rt) = &rp.value {
                                         return self.assign(&lt, &rt, opts.span);
